Fixed issues and included/fixed gH0sTy's code
Using gH0sTy's code, I fixed a pack handle error.
Removed javascript:pingTacker as it was causing a browser error and not loading the URL, meaning the ad kept playing.
Removed the code to overwrite the MOTD file due to base triggers hooking the player typing "motd" and opening the url set in the motd file, with no way to close it after. However, the code always forces the pinion ad to load on connect anyway so it still works fine and allows the server owner to set what MOTD the users sees after typing "motd".